Checking Requirement Error

+++ 0-bubble_sort.c
@@ -11,7 +11,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	int i, j = 0;
+	size_t i, j = 0;
 
 	/*Loop responsible for checking the array again*/
 	for (i = 0; i < size - 1; i++)
